+++ http4k-config/src/main/kotlin/org/http4k/config/Authority.kt
@@ -1,4 +1,4 @@
-package org.http4k.cloudnative.env
+package org.http4k.config
 
 data class Authority(val host: Host, val port: Port? = null) {
     override fun toString() = host.value + (port?.let { ":${it.value}" }.orEmpty())