+++ http4k-config/src/main/kotlin/org/http4k/config/Host.kt
@@ -1,4 +1,4 @@
-package org.http4k.cloudnative.env
+package org.http4k.config
 
 data class Host(val value: String) {
     init {