+++ http4k-config/src/main/kotlin/org/http4k/config/Timeout.kt
@@ -1,4 +1,4 @@
-package org.http4k.cloudnative.env
+package org.http4k.config
 
 import java.time.Duration