+++ http4k-server/shutdown-integration-test/src/main/kotlin/org/http4k/testing/TestServer.kt
@@ -1,8 +1,8 @@
 package org.http4k.testing
 
-import org.http4k.cloudnative.env.Environment
-import org.http4k.cloudnative.env.EnvironmentKey
-import org.http4k.cloudnative.env.enum
+import org.http4k.config.Environment
+import org.http4k.config.EnvironmentKey
+import org.http4k.config.enum
 import org.http4k.server.ServerConfig.StopMode
 import org.http4k.server.asServer
 import org.http4k.testing.TestServerEvent.ServerStarted