+++ src/docs/blog/typesafe_configuration/post/secrets.kt
@@ -1,9 +1,9 @@
 package blog.typesafe_configuration.post
 
 import org.http4k.client.OkHttp
-import org.http4k.cloudnative.env.Environment
-import org.http4k.cloudnative.env.EnvironmentKey
-import org.http4k.cloudnative.env.Secret
+import org.http4k.config.Environment
+import org.http4k.config.EnvironmentKey
+import org.http4k.config.Secret
 import org.http4k.core.Filter
 import org.http4k.core.HttpHandler
 import org.http4k.core.then