+++ src/docs/guide/reference/cloud_native/example_k8s.kt
@@ -3,13 +3,13 @@ package guide.reference.cloud_native
 import org.http4k.client.JavaHttpClient
 import org.http4k.cloudnative.Http4kK8sServer
 import org.http4k.cloudnative.asK8sServer
-import org.http4k.cloudnative.env.Environment
-import org.http4k.cloudnative.env.EnvironmentKey
-import org.http4k.cloudnative.env.Secret
 import org.http4k.cloudnative.health.Completed
 import org.http4k.cloudnative.health.Health
 import org.http4k.cloudnative.health.ReadinessCheck
 import org.http4k.cloudnative.health.ReadinessCheckResult
+import org.http4k.config.Environment
+import org.http4k.config.EnvironmentKey
+import org.http4k.config.Secret
 import org.http4k.core.Filter
 import org.http4k.core.Method.GET
 import org.http4k.core.Request